rename #70 ステートファイルの管理先はReactでもstoreが使われる
