+++ src/component/Recoil/RecoilCounter.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useRecoilState } from "recoil";
-import { counterAtom } from "../../state/atom";
+import { counterAtom } from "../../store/atom";
 
 export default function RecoilCounter() {
   const [counter, setCounter] = useRecoilState<number>(counterAtom);
